refactor: modularise a bit and fix issues
