+++ apis/nucleus/src/components/listbox/assets/get-list-sizes/constants.js
@@ -1,0 +1,4 @@
+export const SCROLL_BAR_WIDTH = 10; // TODO: ignore this - instead set the styling only show on hover...
+export const ITEM_MAX_WIDTH = 150;
+export const FREQUENCY_WIDTH = 40;
+export const ITEM_MIN_WIDTH = 56;
