+++ apis/nucleus/src/components/listbox/assets/get-list-sizes/index.js
@@ -1,0 +1,3 @@
+import getListSizes from './get-list-sizes';
+
+export default getListSizes;
